+++ packages/orq-rc/src/orq_poc_python_multi_env_version/models/createpromptversionop.py
@@ -459,7 +459,6 @@ class CreatePromptVersionMetadata(BaseModel):
 
 
 class CreatePromptVersionRequestBodyTypedDict(TypedDict):
-    id: str
     display_name: str
     prompt_config: CreatePromptVersionPromptConfigTypedDict
     metadata: CreatePromptVersionMetadataTypedDict
@@ -469,8 +468,6 @@ class CreatePromptVersionRequestBodyTypedDict(TypedDict):
 
 
 class CreatePromptVersionRequestBody(BaseModel):
-    id: Annotated[str, pydantic.Field(alias="_id")]
-
     display_name: str
 
     prompt_config: CreatePromptVersionPromptConfig
@@ -515,16 +512,14 @@ class CreatePromptVersionRequestBody(BaseModel):
 
 
 class CreatePromptVersionRequestTypedDict(TypedDict):
-    prompt_id: str
+    id: str
     r"""Prompt ID"""
     request_body: NotRequired[CreatePromptVersionRequestBodyTypedDict]
 
 
 class CreatePromptVersionRequest(BaseModel):
-    prompt_id: Annotated[
-        str,
-        pydantic.Field(alias="id"),
-        FieldMetadata(path=PathParamMetadata(style="simple", explode=False)),
+    id: Annotated[
+        str, FieldMetadata(path=PathParamMetadata(style="simple", explode=False))
     ]
     r"""Prompt ID"""