+++ packages/orq-rc/src/orq_poc_python_multi_env_version/models/getallpromptsop.py
@@ -1413,7 +1413,7 @@ class GetAllPromptsItems3(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = dateutil.parser.isoparse("2024-11-26T03:17:57.484Z")
+    updated: Optional[datetime] = dateutil.parser.isoparse("2024-11-26T10:00:21.083Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")
@@ -2598,7 +2598,7 @@ class GetAllPromptsItems2(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = dateutil.parser.isoparse("2024-11-26T03:17:57.484Z")
+    updated: Optional[datetime] = dateutil.parser.isoparse("2024-11-26T10:00:21.083Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")
@@ -3811,7 +3811,7 @@ class GetAllPromptsItems1(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = dateutil.parser.isoparse("2024-11-26T03:17:57.484Z")
+    updated: Optional[datetime] = dateutil.parser.isoparse("2024-11-26T10:00:21.083Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")