+++ packages/orq-rc/src/orq_poc_python_multi_env_version/models/getallprompttemplatesop.py
@@ -1442,7 +1442,7 @@ class Items3(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = dateutil.parser.isoparse("2024-11-26T03:17:57.484Z")
+    updated: Optional[datetime] = dateutil.parser.isoparse("2024-11-26T10:00:21.083Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")
@@ -2671,7 +2671,7 @@ class Items2(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = dateutil.parser.isoparse("2024-11-26T03:17:57.484Z")
+    updated: Optional[datetime] = dateutil.parser.isoparse("2024-11-26T10:00:21.083Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")
@@ -3917,7 +3917,7 @@ class Items1(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = dateutil.parser.isoparse("2024-11-26T03:17:57.484Z")
+    updated: Optional[datetime] = dateutil.parser.isoparse("2024-11-26T10:00:21.083Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")